+++ gunz/src/PlayerList.cpp
@@ -10,7 +10,7 @@ boost::uint8_t PlayerList::Length() const
 	const size_t realLength = Collection<Player*, std::vector<Player*> >::Length();
 
 	if(realLength <= UCHAR_MAX)
-		return realLength;
+		return static_cast<uint8_t>(realLength);
 	else
 		return UCHAR_MAX;
 }
